Fix unused private fields + split CurriculumVitaeGetterFromExampleXMLFileTest
